# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ariadne Daemon API",
    "description": "REST API of ariadned, the coding-agent orchestrator daemon.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/v1/goals/{id}/complete": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Complete the goal: it moves active -> completed and every session of it\nis torn down.",
        "description": "The orchestrator's call, because it is the only agent that knows whether\nthe plan did what the goal asked for — the daemon can see that every task\nended, and not whether the goal is met. The user's too: it is their goal,\nand a goal whose orchestrator will not start is otherwise one nothing can\nclose.\n\nWhat the daemon checks is the part it can see. A goal with a task still\ngoing is not one anybody may declare finished, however sure they are.",
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "goal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/tasks/{id}/step/complete": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": "step_gate_failed"
          }
        }
      }
    },
    "/v1/tasks/{id}/step/fail": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "fail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "403": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentUsageDto": {
        "type": "object",
        "description": "What one staffed agent spent on a task, named the way a reader addresses\nit: an agent has no name of its own, so its skills are what identify it.",
        "required": [
          "agent_id",
          "skills",
          "usage"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The skills the agent loads; empty only if the agent is gone."
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          },
          "usage": {
            "$ref": "#/components/schemas/TokenUsageDto"
          }
        }
      },
...
      "CompleteGoalRequest": {
        "type": "object",
        "description": "Body of `POST /v1/goals/{id}/complete`: the orchestrator says the goal is\ndone. Its call, not the user's, and it carries nothing — every task being\nfinished or cancelled is the whole of the argument, and the daemon checks\nthat itself.",
        "additionalProperties": false
      },
      "CompleteStepRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "merge_commit": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "FailStepRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ForgeDto": {
        "type": "object",
        "description": "The forge a repository's remote is on, and whether Ariadne works with it.",
        "required": [
          "webhook",
          "kind",
          "host",
          "owner",
          "name",
          "remote",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string",
            "description": "Lower-cased, like `owner` and `name`.",
            "example": "github.com"
          },
          "kind": {
            "$ref": "#/components/schemas/ForgeKind"
          },
          "login": {
            "type": [
              "string",
              "null"
            ],
            "description": "The account the forge CLI is signed in as, stored on enable."
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "remote": {
            "type": "string",
            "description": "The remote it was read off, `origin` where there is one."
          },
          "review_effort": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_model": {
            "type": [
              "string",
              "null"
            ],
            "description": "The pin of the session that reviews a request; null starts none."
          },
          "webhook": {
            "$ref": "#/components/schemas/WebhookDto"
          }
        }
      },
      "ForgeKind": {
        "type": "string",
        "description": "The forge a repository's remote is on (025): which CLI speaks to it.",
        "enum": [
          "github",
          "gitlab"
        ]
      },
...
      "GoalDto": {
        "type": "object",
        "required": [
          "steps",
          "id",
          "title",
          "description",
          "status",
          "orchestrated",
          "model",
          "landing",
          "repos",
          "usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "effort": {
            "type": [
              "string",
              "null"
            ],
            "description": "The reasoning effort that model is run at, pinned like `model`. None =\nwhatever the agent runs it at on its own.",
            "example": "high"
          },
          "id": {
            "type": "string"
          },
          "issue_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "How every task of the goal ends, chosen when the goal was created."
          },
          "model": {
            "type": "string",
            "description": "What the orchestrator or adopted author runs on, `<agent>:<model>`:\nthe registry agent and, after the `:`, the model of it.",
            "example": "claude-acp:claude-opus-5"
          },
          "orchestrated": {
            "type": "boolean",
            "description": "Whether this goal has an orchestrator for its lifetime."
          },
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalRepositoryDto"
            },
            "description": "The registered repositories the goal works in, as they stand now: a\ngoal references them, so an edit to one shows up here."
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            }
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/GoalUsageDto",
            "description": "What the agents of this goal have spent between them."
          },
          "workflow": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GoalRepositoryDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RepositoryDto"
          },
          {
            "type": "object",
            "properties": {
              "goal_branch": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Null until a feature branch goal creates its branch at plan finalization."
              }
            }
          }
        ],
        "description": "A registered repository together with the branch this goal owns in it."
      },
      "GoalStatus": {
        "type": "string",
        "description": "Goal lifecycle status.",
        "enum": [
          "planning",
          "active",
          "completed",
          "cancelled"
        ]
      },
      "GoalUsageDto": {
        "type": "object",
        "description": "What a goal cost, by the seat that spent it. Grouped by seat rather than\nby agent: a goal's authors are as many as it has tasks, and what is\nread at this height is where the tokens went, not which agent went there.",
        "required": [
          "agents",
          "total",
          "orchestrator",
          "authors",
          "reviewers"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentUsageDto"
            }
          },
          "authors": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "Every author session of every task of the goal."
          },
          "orchestrator": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "The orchestrator's sessions, which belong to no task."
          },
          "reviewers": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "Every reviewer session of every task of the goal, all rounds."
          },
          "total": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "Every session of the goal summed, the orchestrator's included."
          }
        }
      },
...
      "Landing": {
        "type": "string",
        "description": "How the tasks of one goal end.\n\nThe one thing about the end of a task the author has to be told, since the\ncommands it runs differ entirely between the endings. The user chooses it\nonce, for the whole goal, when the goal is created, and every task of the\ngoal follows it: some work lands on the base branch, some goes through a\nrequest the author then sees to its merge, and some has nothing to land at\nall — a report filed, a document published, a release cut. Every ending\nreaches [`TaskStatus::Finished`]; landing is one way of getting there\nrather than the meaning of being there.\n\nWhich forge a published request goes to is *not* here: `origin` says\nwhether it is GitHub or GitLab, and asking the remote at landing time\ncannot go stale the way a second copy of the answer would.",
        "enum": [
          "merge",
          "pull_request",
          "none",
          "feature_branch"
        ]
      },
...
      "ModelRank": {
        "type": "string",
        "description": "A user-set rank, independent of discovery and model availability.",
        "enum": [
          "frontier",
          "balanced",
          "fast",
          "local"
        ]
      },
...
      "PermissionMode": {
        "type": "string",
        "description": "How the ACP runtime answers a tool permission request.",
        "enum": [
          "auto",
          "ask",
          "learn",
          "ai"
        ]
      },
...
      "RepositoryDto": {
        "type": "object",
        "required": [
          "id",
          "path",
          "base_branch",
          "permission_mode",
          "default_landing",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "base_branch": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "default_landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "The landing a new goal uses where its request leaves landing out."
          },
          "default_workflow": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "forge": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ForgeDto",
                "description": "The forge its remote is on, or null where the checkout has no usable\nremote (025)."
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Absolute path of the checkout."
          },
          "permission_mode": {
            "$ref": "#/components/schemas/PermissionMode",
            "description": "How the ACP permission requests of every session in this checkout are\nanswered."
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "Seat": {
        "type": "string",
        "description": "Where an agent sits: the orchestrator of a goal, or the author or a\nreviewer of one task.\n\nA seat is a position, not an identity. Every agent below the orchestrator\nis generic, and what it can do comes from the skills it loads; the seat is\nonly what the state machine and the launcher need to know about it.",
        "enum": [
          "agent",
          "orchestrator",
          "author",
          "reviewer"
        ]
      },
...
      "StepGate": {
        "type": "string",
        "description": "What a step waits for before the next one may start: the author's commit,\na push, a merge onto the base, or a request merged by a human.",
        "enum": [
          "committed",
          "pushed",
          "merged",
          "request_merged"
        ]
      },
...
      "TaskAgentDto": {
        "type": "object",
        "required": [
          "id",
          "seat",
          "skills",
          "model"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "The branch this agent works on: the task branch for the first author,\na suffixed sibling of it for every later one. None for a reviewer,\nwhich owns no branch."
          },
          "brief": {
            "type": [
              "string",
              "null"
            ],
            "description": "What the orchestrator told this agent beyond the task itself. None =\nthe task is the whole of it."
          },
          "effort": {
            "type": [
              "string",
              "null"
            ],
            "description": "The reasoning effort that model is run at. None = whatever the agent\nruns it at on its own.",
            "example": "high"
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "description": "What this agent runs on, `<agent>:<model>`.",
            "example": "codex-acp:o3"
          },
          "seat": {
            "$ref": "#/components/schemas/Seat",
            "description": "`author` or `reviewer`."
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "This agent's live session, the id `POST /v1/sessions/{id}/switch`\ntakes. None while it carries no live session."
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The skills this agent loads, in the order they reach it.",
            "example": [
              "coding",
              "documentation"
            ]
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TaskDto": {
        "type": "object",
        "required": [
          "id",
          "goal_id",
          "repo_id",
          "title",
          "description",
          "status",
          "agents",
          "depends_on",
          "branch",
          "landing",
          "stalled",
          "picks",
          "usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskAgentDto"
            },
            "description": "The agents staffed on the task: the authors first, then the reviewers\nin review order. What each one can do is the skills it carries. Most\ntasks staff one author; one staffed with several runs them in\nparallel, and the reviewers pick the branch that lands."
          },
          "branch": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Ids of tasks that must merge before this one starts."
          },
          "description": {
            "type": "string"
          },
          "goal_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "How the task ends: a read-only copy of its goal's landing, which every\ntask of the goal follows."
          },
          "merge_commit": {
            "type": [
              "string",
              "null"
            ]
          },
          "picked_agent_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The author the reviewers picked, on a task staffed with several: the\none whose branch lands. None for a one-author task, and until the\npick settles."
          },
          "picks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskPickDto"
            },
            "description": "The picks the reviewers have recorded so far, oldest first. Empty for\na one-author task."
          },
          "pr_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL of the pull or merge request the task was published as, once its\nauthor has reported one; None for a task landed directly."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why a `failed` or `cancelled` task ended — the author's own\n`fail_task` reason, a dependency that never landed, a cancelled goal.\nNone for every other status, and for an ending nobody gave a reason\nfor."
          },
          "repo_id": {
            "type": "string",
            "description": "Id of the repository the task works in, one of its goal's."
          },
          "stalled": {
            "type": "boolean",
            "description": "Set when the agent went idle without advancing the task."
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/TaskUsageDto",
            "description": "What the agents of this task have spent between them."
          },
          "worktree_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TaskPickDto": {
        "type": "object",
        "description": "One agent staffed on a task: where it sits, what it knows, and what it\nruns on.\n\nThe agent has no identity of its own. `seat` says only whether it authors\nthe task or reviews it; the skills are what it can do. What it runs on was\nsized by the orchestrator when it staffed the task, or chosen by the user\nsince — either way it is what this agent runs on, and nothing behind it\nchanges that.\nOne reviewer's pick of the winning author, on a task staffed with several\nauthors.",
        "required": [
          "reviewer_agent_id",
          "author_agent_id",
          "created_at"
        ],
        "properties": {
          "author_agent_id": {
            "type": "string",
            "description": "The author it picked."
          },
          "created_at": {
            "type": "string"
          },
          "reviewer_agent_id": {
            "type": "string",
            "description": "The reviewer that picked. One pick per reviewer per task."
          }
        }
      },
      "TaskStatus": {
        "type": "string",
        "description": "Task lifecycle status.",
        "enum": [
          "pending",
          "ready",
          "in_progress",
          "under_review",
          "changes_requested",
          "approved",
          "finished",
          "cancelled",
          "failed"
        ]
      },
...
      "TaskUsageDto": {
        "type": "object",
        "description": "What a task cost, by who spent it: its author, its reviewers one entry\neach, and the total of every session on the task.",
        "required": [
          "agents",
          "total",
          "author",
          "reviewers"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentUsageDto"
            }
          },
          "author": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "The author's own, across every run of it."
          },
          "reviewers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentUsageDto"
            },
            "description": "One entry per reviewer that has a session on the task, every review\nround of it summed, in review order. A reviewer whose session has yet\nto report anything is listed with zeros; one that has never been\nspawned is not listed at all."
          },
          "total": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "Every session on the task summed, whatever its seat."
          }
        }
      },
...
      "TokenUsageDto": {
        "type": "object",
        "description": "Tokens spent, as the agents' own transcripts report them.\n\nAlways present and always a number: nothing reported is zero, not null.",
        "required": [
          "input_tokens",
          "cached_input_tokens",
          "output_tokens"
        ],
        "properties": {
          "cached_input_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "The subset of `input_tokens` served from the prompt cache, so never\nadded to it.",
            "minimum": 0
          },
          "input_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "Prompt tokens, cache reads and cache writes included.",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "Completion tokens, thinking and reasoning included.",
            "minimum": 0
          }
        }
      },
...
      "WebhookDto": {
        "type": "object",
        "description": "Public hook status. The hook secret is never serialized.",
        "required": [
          "state"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "fetch_error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the last fetch of the repository's requests failed, or null once\none works: what says polling works, where no hook is live."
          },
          "last_delivery_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WorkflowStepDto": {
        "type": "object",
        "description": "One column of a workflow.",
        "required": [
          "id",
          "title",
          "description",
          "skills"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "gate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StepGate",
                "description": "What the step waits for before the next one starts, where the\ndocument named one."
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "Kebab-case, unique within the workflow.",
            "example": "develop"
          },
          "rank": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModelRank",
                "description": "The model rank the step prefers, where the document named one."
              }
            ]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The skills an agent on this step loads."
          },
          "title": {
            "type": "string",
            "description": "The display title.",
            "example": "Develop"
          }
        }
      }
    }
  },
...
}